DERBY-5223 Thread's interrupted flag not always preserved after Derby returns from JDBC API call
Patch DERBY-5223b: This fix moves the initialization of the variable
"interruptedException" earlier in
GenericLanguageConnectionContext#initialize and adds a missing
reinitialization to resetFromPool as a precaution (if
interruptedException is still non-null the connection should have
throws 08000).

The patch also changes the InterruptResilienceTest so that JUnit
asserts in the worker threads will get propagated to the main test
thread on completion, so any future errors in these invariants do not
get overlooked.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1102826 13f79535-47bb-0310-9956-ffa450edef68