DERBY-3870: Concurrent Inserts of rows with XML data results in an exception
Remove lazy initialization of the field that holds the cached
SqlXmlUtil instance. This simplifies the generated byte code. It also
removes the need for an explicit syntax check of the XML query during
the bind phase, as the earlier initialization ensures that syntax
errors will be detected at compile time.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1126358 13f79535-47bb-0310-9956-ffa450edef68