DERBY-5251: make ErrorCodeTest pass in non-English locale
This change was contributed by Houx Zhang (houx.zhang at gmail.com)

This change modifies ErrorCodeTest to use LocaleTestSetup to
force the test to run in English locale. Since the point of
this test is to test the behavior of syscs_diag.error_messages,
by comparing the contents of the messages to hard-coded English
strings, the test needs to run in the matching locale.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1128942 13f79535-47bb-0310-9956-ffa450edef68