DERBY-5243 assert failure in test testRAFReadWriteMultipleThreads: interrupted flag cleared
Patch DERBY-5243-1: It moves the setting of the interrupted flag to
just before CONN_INTERRUPT exception is thrown as an SQLException back
to the user application, i.e. as late as possible. Previously, we set
the interrupted flag when we first threw the internal
StandardException. The new placement means that we are done with IO to
derby.log which may be behind the problem we are seeing on older
VMs. I ran the new version 100 times on Linux/JDK 5 without seeing the
error, which previously appeared in ca 20% of the runs.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1129797 13f79535-47bb-0310-9956-ffa450edef68