DERBY-1482/DERBY-5121
Rick Hillegas contributed a trigger test for DERBY-1482/DERBY-5121. With revision 1125453, that test was added to Changes10_8 but this test really is applicable for upgrades from all releases and should not be added into a specific version upgrade test. As a result, I am moving the test from Changes10_8.java to BasicSetup.java. This will ensure that the trigger test will get run for upgrades from all previous releases.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1130895 13f79535-47bb-0310-9956-ffa450edef68