DERBY-3870: Concurrent Inserts of rows with XML data results in an exception
Allow databases that contain triggers with XML operators to be
upgraded from a version without the fix for this issue to a fixed
version. It used to fail because it couldn't deserialize the old plan
for the trigger. Fix it by not trying to deserialize the plan before
invalidating it.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1132546 13f79535-47bb-0310-9956-ffa450edef68