DERBY-5258 btree post commit releases latch before committing/aborting purges, possibly allowing other operation on page
backporting change 1132711 from trunk to 10.8 branch.

Fixing row level btree postcommit reclaim space to hold latch until end
of the internal transaction. Before this fix there was a very small window
(a few instructions) between the release of the latch and the commit of the
transaction where another transaction could access the page, insert rows,
and if a crash happens cause the undo of the purges of the reclaim space work
to fail.

It is proposed that this is what caused DERBY-5248, but without a repro it
can't be proved. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.8@1133470 13f79535-47bb-0310-9956-ffa450edef68