DERBY-1482
Backporting revision 1132664 from trunk to 10.7 codeline. In addition to the backport, had to make couple changes by hand. Added 10.7.1.1 to the list of old revisions of Derby from which there will be upgrade to current 10.7 codeline. The reason for adding 10.7.1.1 is DERBY-5121 which caused data corruption is this release. This was fixed in the 10.7 codeline after 10.7.1.1 and we want to make sure that the faulty triggers created in 10.7.1.1 get fixed when taken to 10.7 codeline. 

Also, normally, when an older release is hard upgraded to the codeline, the database can't be booted again with the old release. But when we are testing upgrade from 10.7.1.1 to current codeline on 10.7, we can always goback to 10.7.1.1 because it is simply a point release upgrade. I have changed the upgrade test to allow this going back to 10.7.1.1 after the db has been hard upgraded to 10.7 codeline.

Commit comments for 1132664
Giving the upgrade tests in Changes10_8 a little more thought, I feel that the more appropriate place for the tests will be in upgradeTests.BasicSetup. The reason for this is that the tests are creating triggers and testing the results when they get fired. This kind of testing is not unique to any particular Derby release and hence it is better for them to go in BasicSetup to make sure that the results are consistent between various phases of upgrades. We do already know that because of the bug DERBY-5121, the results will be incorrect in 10.7.1.1 and the test is verifying those incorrect reulsts for 10.7.1.1. Based on this, I am putting these set of trigger tests in BasicSetup.java This makes Changes10_8.java to have no tests left in it and hence we can delete Changes10_8.java from svn.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.7@1133495 13f79535-47bb-0310-9956-ffa450edef68