DERBY-5252: make GrantRevokeTest pass in non-English locale
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

This change adjusts 4 places in the test to compare the actual
error message text only if the test is running in the English locale.

This enables the test to pass when run in non-English locales, while
also preserving the error message text validation for runs of the
test in the English locale.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1134139 13f79535-47bb-0310-9956-ffa450edef68