DERBY-5289 Unable to boot 10.5.1.1 database - fails during soft/hard upgrade process for a new version number while trying to drop jdbc metadata
merged 1140757 from 10.5 branch. Some manual merge for tests required.
For 10.3 the fix for this issue is the relevant part of the  DERBY-3870 fix contributed by Knut and separated out by Mamta.  The test fix for DERBY-5105 contributed by Knut and the testcase based on the script contributed by Brett Mason.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@1140812 13f79535-47bb-0310-9956-ffa450edef68