DERBY-5292 SQLAuthorisation and views
Patch derby5292d: For views, the premissions collection is disabled
from tables in the query from-list since it should run with definer's
rights. However, this disabling did not work for all cases. The way to
reach all the "from" tables has been changed to use a node visitor
instead, so as to be able to also reach tables inside subqueries and
inside explicit JOIN and set operations, cf the repro issues. An added
test case was added to GrantRevokeTest: testViewDefinersRights.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1142635 13f79535-47bb-0310-9956-ffa450edef68