DERBY-5311: Convert tools/ij4.sql to junit
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

This patch adds a simple JUnit test case which extends ScriptTestCase
and runs ij4.sql. Note that in this situation, we are retaining ij4.sql
and ij4.out, even though we now run them only via JUnit, because we
are testing the behavior of the ij tool itself, and the way it runs
scripts, rather than testing the underlying Derby engine behaviors.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1143062 13f79535-47bb-0310-9956-ffa450edef68