DERBY-5311: Convert tools/ij4.sql to junit
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

This patch moves the execution of ij4.sql from ij4Test.java to
ToolScripts.java. Improvements in the test harness, and in ij itself,
mean that we no longer need a special test case setup for this script,
so ij4Test.java is deleted from the test tree.

The patch also includes a line to reset maximumdisplaywidth in ij
back to the default setting of 128 at the end of the test, so the
test doesn't disrupt the running of subsequent ij tests.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1143416 13f79535-47bb-0310-9956-ffa450edef68