DERBY-5303: Convert tools/URLCheck.sql to junit
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

This patch moves the execution of URLCheck.sql from the old derbytools
suite to the ToolScripts JUnit test suite.

The patch includes a modification to the LocalizedResource class in
the tools.jar library to ensure that if getNewOutput is called, then
subsequent uses of OutputWriter send their output to that output stream.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1144123 13f79535-47bb-0310-9956-ffa450edef68