DERBY-5333 Intermittent assert failure in testInterruptShutdown: thread's interrupted flag lost after shutdown
Patch derby-5333a fixes two instances where we neglected to note that
an interrupt had occured while in a wait call. They were missed during
the first code inspection because they appeared to be part of daemon
code from a superficial inspection, but the code was in fact called by
session threads (for which we expect interrupts).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1151101 13f79535-47bb-0310-9956-ffa450edef68