DERBY-5342: make ScriptTestCase support ij properties
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

This patch refactors the ij utilMain class so that the code
that supports the properties:

	- ij.showNoConnectionsAtStart
	- ij.showNoCountForSelect

is extracted into a separate method so that it can be called
from goScript. In addition, ScriptTestCase is modified so that
it sets these properties to 'true' (other test cases may
subsequently set them to 'false' for testing purposes).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1151691 13f79535-47bb-0310-9956-ffa450edef68