DERBY-5342: make ScriptTestCase support ij properties.
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

This patch fixes a bug in SystemPropertyTestSetup. In
SystemPropertyTestSetup.setUp() setProperties() is called, but a property
won't be added into oldValues if the old value equals the new one, yet then
the property will be removed in tearDown() later. 

The patch revises the behavior so that the property is not removed in this
case.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1152371 13f79535-47bb-0310-9956-ffa450edef68