DERBY-5327: Convert tools/ij5.sql to junit
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

This change converts the ij5.sql test to the JUnit tools suite. The test
verifies that the ij properties that control displaying the list of
connections at startup, and showing the number of rows selected in a
query, are working properly.

The new ij5.java test is added to the tools JUnit suite, and the old
test is removed from derbytools.runall.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1152796 13f79535-47bb-0310-9956-ffa450edef68