DERBY-5271: Client may hang if the server crashes due to a java.lang.Error
Tries to ensure that if the network server crashes due to a condition
raising java.lang.Error, the client socket will be closed on the server
side. Note that even if one of the worker threads crashes, the network
server itself may remain operational. If the JVM process dies, the
sockets will be closed anyway.

Patch file: derby-5271-1a-inital_fix_proposal.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1158108 13f79535-47bb-0310-9956-ffa450edef68