DERBY-5367: Stale data retrieved when using new collation=TERRITORY_BASED:PRIMARY feature
Deoptimize code path for BTree insert when updating columns with a collation
different from UCS BASIC. Simply undeleting the existing matching row (marked
as deleted) may be incorrect, because the value stored there can be different
from the key value used for lookup due to the collation.

Added code to track whether a conglomerate contains a collated column or not,
such that the right insert code path can be chosen. The array of collation ids
is scanned when a conglomerate is created, and when a conglomerate is
"restored" from disk (i.e. readExternal).

Added a test for the new code path (based on the issue report).

Patch file: derby-5367-4c-fix_with_optimization_improved.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1174436 13f79535-47bb-0310-9956-ffa450edef68