DERBY-5363 Tighten default permissions of DB files with >= JDK6
Patch derby-5363-full-5 implements the ability to restrict file
permission of newly created directories and files beyond the default
access (cf. umask on Posix file systems and similar on NTFS), i.e. to
the account creating the file.  This behavior is controlled by a
property, "derby.storage.useDefaultFilePermissions", cf the release
notes attached to issue.  By default the property is true, i.e. gives
the existing (lax) behavior on embedded and with the network server if started
via the API.  If the server is started from the command line, the
new restrictive permissions apply by default.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1176591 13f79535-47bb-0310-9956-ffa450edef68