DERBY-5420 Regression suite appears locale sensitive: failed in TableLockBasicTest: bug in RealBasicNoPutResultSetStatistics
Patch derby-5420-2 changes the way the localization of the floating
point numbers is done for RealBasicNoPutResultSetStatistics to use an
explicit decimal format in the localized message text itself. When the
double argument is henceforth filled in, the locale is already set
correctly (in MessageBuilder), so the the decimal point/comma will be
chosen correctly according to locale.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1176633 13f79535-47bb-0310-9956-ffa450edef68