DERBY-5363 Tighten default permissions of DB files with >= JDK6
Patch derby-5363-followup, which adds a missing accessController block
around setting the system property
SERVER_STARTED_FROM_CMD_LINE. Without the patch, this would fail if
running with a security manager specified on the command
line. 

If the property permission is missing, the error is printed
unconditionally and exit(1) from main is taken. Cf. DERBY-5413 which
tried another (aborted) approach to make sure it got printed.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1177718 13f79535-47bb-0310-9956-ffa450edef68