DERBY-5363 Tighten permissions of DB files to owner with >= JDK7
Patch derby-5363-followup-linux.

RestrictiveFilePermissionsTest for this feature broke on some
platforms (thanks to Kathey for noticing). Apparently, the ACL view of
Posix file system permissions is not available for all Unix/Linux
versions in JDK 1.7 (I had tested on Solaris 11 and Windows). The
changes in the test now fall back on using
PosixFileAttributeView#readAttributes if the ACL view is not
available.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1179042 13f79535-47bb-0310-9956-ffa450edef68