DERBY-5363 Tighten permissions of DB files to owner with >= JDK7
Patch derby-5363-followup-unix.

It turns out there is no guarantee the the underlying file system
supports ACLs even though Files#getFileAttributeView called with
aclFileAttributeViewClz.class as an argument returns an object. We
also need to call the method:

     FileStore#supportsFileAttributeView(AclFileAttributeView.class)

to ascertain whether we have support for ACLs. To get at the current
FileStore, we need to inquire about that given a path:

     Files.getFileStore(<path>)

which requires the RuntimePermission "getFileStoreAttributes", hence
the current patch's changes to the policy files.

With the patch, RestrictiveFilePermissionsTest run OK on Solaris/UFS. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1180713 13f79535-47bb-0310-9956-ffa450edef68