DERBY-5447: Deadlock in AutomaticIndexStatisticsTest.testShutdownWhileScanningThenDelete (BasePage.releaseExclusive and Observable.deleteObserver (BaseContainerHandle))
Clean the daemon context only after the running worker thread (if any) has
finished to avoid Java deadlock when closing the container handles obtained
with the context.
The deadlock is intermittent, but can easily be reproduced, and involves
synchronization in BasePage and in java.util.Observable.

Patch file: derby-5447-2a-change_istat_shutdown.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1180790 13f79535-47bb-0310-9956-ffa450edef68