DERBY-5406: Intermittent failures in CompressTableTest and TruncateTableTest
If GenericActivationHolder determines that a recompile is needed, it now
throws an exception to signal that to the caller instead of doing the
recompilation itself. This way, if the statement is invalidated again
during the recompilation, the already existing retry logic in the caller
(that is, GenericPreparedStatement.executeStmt()) will be used to detect
that the recompilation must be retried.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1189067 13f79535-47bb-0310-9956-ffa450edef68