Backporting revision 1203252 from trunk to 10.8 codeline. Commit comments were as follows
DERBY-5379 testDERBY5120NumRowsInSydependsForTrigger - The number of values assigned is not the same as the number of specified or implied columns. 
DERBY-5484 Upgradetest fails with upgrade from 10.8.2.2 (7 errors, 1 failure) on trunk 

The above 2 jiras are duplicates. The upgrade tests are failing when doing an upgrade from 10.8.2.2 to trunk. 

The tests that are failing were written for DERBY-5120, DERBY-5044. Both these bugs got fixed in 10.8.2.2 and higher. 

The purpose of these tests is to show that when the tests are done with a release with those fixes missing, we will see the incorrect behavior but once the database is upgraded to 10.8.2.2 and higher, the tests will start functioning correctly. The problem is that we do not recognize that if the database is created with 10.8.2.2, then we will not the problem behavior because 10.8.2.2 already has the required fixes in it for DERBY-5120 and DERBY-5044. I have fixed this by making the upgrade test understand that incorrect behavior would be seen only for releases under 10.8.2.2 




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.8@1203799 13f79535-47bb-0310-9956-ffa450edef68