DERBY-5459 Result set metadata are out of sync on client after underlying table is altered
Patch derby-5459-3. We now resend the result set metadata to the
client when the cursor is opened if the prepared statement gets
recompiled due to it being out of date when the server tries to
execute it (DRDAConnThread line 871). To detect this we introduce a
version counter which is incremented each time a statement is
(re)compiled and make a note which version's metadata gets sent to the
client as part of the explicit prepare. That version is compared with
the current version when we execute to make the decision whether to
resend metadata or not.

This also fixes DERBY-2402, a duplicate.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1205426 13f79535-47bb-0310-9956-ffa450edef68