DERBY-5559 AssertFailures (7, or 8) with ibm 1.6 and 1.5 on Windows XP in lang.NativeAuthProcs fixture testAll
This change just adds explicit printout in the tests when the asserts fail,
it does not fix the test failures.
I added them to the 4 places I saw them failing in nightly runs as reported
in the apache nightly test run site for the following report:
http://people.apache.org/~myrnavl/derby_test_results/main/windows/testlog/ibm15/1227449-suites.All_diff.txt

All the problem areas are asserts based on timestamps and the tests seem to 
be waiting for some time such that these asserts will be true.  My guess is
that this methodology is not portable with the current values and may need to
be tuned more to work across all JVMs and hardware. 

I am seeing errors on a XP laptop running ibm jvm 16.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1228537 13f79535-47bb-0310-9956-ffa450edef68