DERBY-5559 AssertFailures (7, or 8) with ibm 1.6 and 1.5 on Windows XP in lang.NativeAuthProcs fixture testAll
Adding longer waits and some additional waits to fix problems on my laptop with
this test.  Before this change the test would expect the timestamp to have
moved forward after some password manipulation operations, but instead the
timestamp would be the same as before.  I believe this is just a problem with
the known issue of JVM/machine caching of time, and not an actual problem
with the password mechanism.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1229835 13f79535-47bb-0310-9956-ffa450edef68