DERBY-5564 Code does different things depending if derby.locks.deadlockTrace=true is set
backporting change #1230100 from trunk to 10.5 branch.  Automatic merge of
changes was clean.

original patch by Brett Bergquist, then modified by Mike Matrigali for submission.

Changes the code to always return 40XL1 as the SQL state when a lock timeout
occurs. Previous to this change if deadlock diagnostics were enabled then
40X02 would be returned. Internally multiple places in the code was not
expecting the second error code for a lock timeout. Also it was agreed that
it was confusing for user applications to have to code for both states in
case they wanted to turn diagnostics on and off.

Existing test cases were changed to match the new expected behavior.

The behavior in DDLConstantaction to immediately throw an error on first try
if a lock timeout is encounted with diagnostics enabled was preserved. The
    error thrown now will be with sql state 40XL1 and not 40X02 as before.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@1231797 13f79535-47bb-0310-9956-ffa450edef68