DERBY-4540 AssertionFailedError in store.AccessTest.testCS4595B_UniqueIndex(AccessTest.java:1729)' on SUSE Linux / IBM JIT - r9
backported change #1233098 from trunk to 10.8 branch, merged cleanly.

This change just adds some diagnostic print out to the assert that seems
to be intermittently failing.  Next time we should see the number of pages
expected and the full dump of the statistics output the test is getting when
it gets this failure.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.8@1233108 13f79535-47bb-0310-9956-ffa450edef68