DERBY-5567 AlterTableTest#testDropColumn fails: drop view cannot be performed due to dependency
Backported from trunk with some minor edits (imports).

When a view (b) is defined on another view (a), dropping a column in
the base table can lead to both view being invalidated. This patch
(DERBY-5567-1) fixes a problem in the logic:

Depending on the order in which dependencies of the base table column
are registered, the invalidation will happen either view a or view b.
If it happens on view a first, this view in turn will try to
invalidate view b (since that depends on view a), but with the
DROP_VIEW action which fails. The patch changes this recursive
invalidation to use the original action, e.g. DROP_COLUMN which will
allow dropping the dependent view (dropping a column is allowed to
cause cascading drops of dependent views). 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.8@1240258 13f79535-47bb-0310-9956-ffa450edef68