DERBY-3740 BlobClob4BlobTest.testLockingWithLongRowBlob fails with .AssertionFailedError: FAIL - statement should timeout
The 2 tests that were failing intermittently (testLockingWithLongRowBlob 
and testLockingBlob) would both select a target blob known to be returned
as a stream, and assign it to a variable.  The expectation was that until
the end of the routine Derby would maintain a read lock on the blob.  The
problem was that it looks like some JVMs/JITs would recognize that the variable
was never used and call finalize before the test executed the code to 
get an expected lock timeout on updating the blob in another connection.
Fixed by adding references to the variable in code executed after the 
expected timeout.  Before the fix test was failing about 1/2 time in my
environment, after the fix the test ran 100 times without a failure.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1240521 13f79535-47bb-0310-9956-ffa450edef68