DERBY-3823 NullPointerException in stress.multi test
backported changes #1183192 and #1182570 from trunk to 10.8

This patch fixes a race condition in
EmbedPreparedStatement#getMetaData: if we are trying to retrieve the
metadata for a prepared statement while it is being recompiled there
is a time window during which the activation class is null. The
existing code could therefore cause an NPE. The new code plugs the
race condition. This NPE led to intermittent errors in
stressTestMulti. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.8@1242007 13f79535-47bb-0310-9956-ffa450edef68