DERBY-5599 readlocks.sql fails with extra locks.
backported #1245056 from trunk to 10.6 branch.

readlocks checks for expected locks in a number of scenarios. Intermittently
it could see extra locks if it happened to check the lock table, while the
background thread was running reclaiming deleted rows. This is most likely
in the tests that put a single row per page, delete some rows, commit and
then shortly after do work and check the lock table.

Added workaround used other places which waits for post commit to finish
after the commit following the deletes. Have updated all the places I saw
where the problem is likely - not just the one causing the diff reported
in this JIRA.

Changes affect the readlocks test and the updatelocksJDBC30 test. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@1291337 13f79535-47bb-0310-9956-ffa450edef68