DERBY-5631: Extend SecurityManagerSetup to add extra privileges to the set of default privileges (merge two policy files)
Added the capability to merge two policy files in SecurityManagerSetup. For now
the only merge allowed is to merge an additional policy resource with the
default test policy. The intended use-case is for tests requiring a few extra
permissions to be runnable in the test framework.
Refreshed class comment.
Made getDefaultPolicy() private.
Added a debug println stating which policy is installed by the decorator.

Patch file: derby-5631-1c-merge_policy_files.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1295436 13f79535-47bb-0310-9956-ffa450edef68