DERBY-2031: Convert derbynet/testProtocol.java to JUnit
Rewrote testProtocol to JUnit, now called ProtocolTest. The tests themselves
still live in a separate file.
Added the new adapter ProtocolTestAdapter, which lives in the impl.drda
package. Added this file to the list of extra files to include in derbynet.jar.
Enabled the new test as part of derbynet._Suite and thus suites.All.
The test is compiled to JDK 1.5 level, and doesn't run on J2ME.

Patch file: derby-2031-5b-new_adapter.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1297517 13f79535-47bb-0310-9956-ffa450edef68