DERBY-5638 intermittent test failure in test_05_ClobNegative when running full largedata._Suite; LobLimitsTestjava.sql.SQLException: Table/View 'BLOBTBL' already exists in Schema 'APP'.
The test fixture test_05_ClobNegative which runs into intermittent test failures does not commit/rollback the transaction after doing a series of negative testing. This may be the cause of left over locks which does not allow CleanDatabaseSetup to get the required locks to drop the tables as part of it's cleanup. As part of this change, we are adding a rollback at the end of test_05_ClobNegative method and see if that helps with the intermittent locking issues.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1298907 13f79535-47bb-0310-9956-ffa450edef68